# Conflicts:
#	apps/wallet/instance/app/src/main/java/com/tonapps/tonkeeper/koin/KoinModule.kt
#	apps/wallet/instance/app/src/main/java/com/tonapps/tonkeeper/ui/screen/root/RootViewModel.kt
#	settings.gradle.kts

+++ tonapi/src/main/kotlin/io/tonapi/models/TraceID.kt
@@ -33,7 +33,7 @@ data class TraceID (
     val id: kotlin.String,
 
     @Json(name = "utime")
-    val utime: kotlin.Int
+    val utime: kotlin.Long
 
 )
